fix markdown content not displaying
